Update README.md
Data

The Number of Death Related to State Conflict have been obtained from Uppsala Data Conflict Program
(https://ucdp.uu.se/ )

The rest of the data (coming from various databases) have been formatted to a uniform style and offered online by “Our World in Data” (https://ourworldindata.org/ ). 

Model

Ensemble Optimal Interpolation

Here we employ a relatively simple data-driven model of the state evolution using public-domain data (see details below in “State indicators”) and the machinery of the Ensemble Optimal Interpolation   (Evensen, 2003; Francçois Counillon & Laurent Bertino, 2009) We define each country through a collection of state indicators such as area, population, GDP, CO2 emissions, inequality, life expectancy, among others. These indicators collectively characterize a country and often vary over time. To analyse these variations, we assume a multivariate Gaussian distribution of these indicators evolving through time according to the forecast model:

ψ_t^f= ψ_(t-1)^a  =ψ_t^#+ε 									(1)	
Here   ψ_t^f  is a vector of state indicators predicted by the model at time t, ψ_t^# is a true value of the state vector at time t, ψ_(t-1)^a is the best estimate of the state available from the previous time step (called analysis), and ε is a multivariate Gaussian process with zero mean μ=0 and the covariance matrix is inferred from the ensemble trajectories of the historical data ‘A’ spanning from 1970 to 2017:
P=(ε ε^T ) ̅≈1/(N-1) (A-A ̅ ) (A-A ̅ )^T 							(2)
Here N=195 is the size of the ensemble where each member represents a particular country. The number of state variables defining a country is 32. To ameliorate the impact of the nonstationary nature of the state records, prior to using (2), the data is normalised and then integrated globally to calculate mean values for each year separately. From these annual means, we estimate the error-covariance matrices yearly and then integrate them over the time to produce a mean covariance applicable across the years (referred to as background error-covariance matrix).
Analogous to the model, observations are assumed multivariate Gaussian with the known mean and error covariance matrix:
d_t^ =Hψ_t^ +ξ 								(3)
R=(((ξ-ξ ̅ ) (ξ-ξ ̅ )^T ) ) ̅ 							(4)
Here d is a vector of observations, H is an observation matrix, and ξ  is an observation error given by Gaussian distribution with zero mean and a diagonal covariance matrix R. The entries into R matrix are inferred from the model calibration runs. 

During the assimilation run predictions by the model are updated every simulation time step by assimilating observations (3) into the model (1): 
ψ_t^a=ψ_t^f+K(d_t^ -Hψ_t^f ) 							(6)
〖K=PH〗_^T (HPH^T+R)^(-1) 							(7)
Equations (1-7) complete the description of the data-assimilating model. During the simulation every state is represented by an ensemble with the varying number of members, each member corresponding to a random sample from the distribution of  ε.

Gaussian distribution of errors conditioned on perturbations
To run scenarios with perturbation experiments beyond the time period covered by observations, the model state is updated by calculating the mean of the Gaussian distribution of errors conditioned on known perturbations (rather than by assimilating data). To this end, the model error ε  is partitioned into the unknown errors ε1 and the known perturbations  ε2 
ε=〖[ε1,ε2]〗^T   									(7)
Similar partitioning is applied to the mean of the error μ and the error-covariance matrix P:
μ=〖[μ1,μ2]〗^T   									(8)
[■(P_11&P_12@P_21&P_22 )] 									(9)
The conditional distribution of the first partition ε1  given the known perturbation ε2=δ is a multivariate Gaussian with the mean
(ε1) ̅  = μ1+P_12   P_22^(-1)  (δ- μ2)  							(10)
and the covariance matrix
( P_11 ) ̅  = P_11  - P_12   P_22^(-1)  P_21  							(11)
Since the mean of errors in (1) is assumed zero, equation (10) reduces to 
 (ε1) ̅  = P_12   P_22^(-1)  δ  								(12)
For a given vector of perturbations δ we can now evaluate the remaining errors via (12), and then propagate the state vector forward in time via model (1).